requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
Django==1.10.2

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==3.2.2

# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.7.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.5.0
redis==2.10.5
celery==3.1.24

# Your custom requirements go here
django-taggit==0.21.3
dropbox==6.9.0
django-pure-pagination==0.3.0
django-filter==0.15.2
djangorestframework==3.4.7
